refactor: move leaderboard endpoints to a dedicated router for cleaner route organization
